Refactor TreeBuilder to implement TreeGeneratorScope
This change introduces a new interface `TreeGeneratorScope`, allowing both `TreeBuilder` & `ChildrenGeneratorScope` to implement unified set of methods. This ensures safer and generalized way to add and generate nodes and leaves. Also, it removed unnecessary `isOpen` attribute from `Node` class as the state should be managed by client, not by the model class. This commit helps in standardizing the way elements are added and generated across classes.