Refactor theme generator gradle plugin and optimize sarif results (#95)
* "Refactor theme generator and optimize sarif results"

This commit includes the optimization and refactor of the codebase. In the build.gradle files, the 'intUiLight' and 'intUiDark' themes have been registered and their properties have been updated. The Theme Generation plugin has been changed to use project conventions and the color group reading process was simplified. There is an optimized caching strategy particularly for SVG loading. Sarif results are now generated with kotlinx.sarif, which allows fluent API for handling the sarif format. Also, it includes switching from the java-sarif library to kotlin-sarif.

The changes enhance the robustness and reliability of the code. They also make it easier to understand, modify, and maintain the code. And using Kotlin-based libraries optimizes the performance and simplifies future changes in the project.

* Update build config for cleaner syntax

The build config file has been updated to use a cleaner syntax. The redundant use of .set() has been replaced with direct assignment for better readability. Also, added a suppression for UnstableApiUsage warning at the top of the files.