Remove unnecessary coroutine and state updates in SelectableLazyColumn
The previous commit included the removal of an unnecessary coroutine that was launched for scrolling to the previously active item in the SelectableLazyColumn. Also, certain state updates associated to last active item index and selected keys were taken out as part of cleaning up unnecessary statements. Changes were made to increase efficiency of handling list navigation and selection in the SelectableLazyColumn. These lines of codes were found to be extraneous according to the newer logic regarding layout handling.