Fabrizio.scarponi/111 spinning indicator circular progress indicator (#134)
* wip

* wip

* Add a customizable circular progress to the UI

This commit adds a flexible circular progress option to the UI that supports both small and large size variants. The base theme has been updated to provide styling to both variants, and new SVG files have been added to provide dynamic views of the circular progress. The circular progress can be incorporated in the UI using the CircularProgress and CircularProgressBig components. This addition provides a more visually engaging way to show progress in the UI.

* Refactor names of circular progress components and styles

* wip

* wip

* Refactor circular progress styles

* Remove unused spinner SVG files

* Remove SvgLoader from CircularProgressStyle

* Remove SvgLoader from CircularProgressStyle

* Improve handling of non-existent colors

Added a `FallbackMarker` in BridgeUtils.kt, improving the named color retrieval process. The `retrieveColorOrNull` method now checks for colors that are not found and returns a specific marker, rather than an inappropriate color. This avoids any confusion and enhances the overall color management within the application."

* Update svgLoader usage for progress bars

* fixes