Complete Swing demo
Had to give up on the image loading at some point, it was just a PITA.
I wanted to also have a placeholder, and less terrible performance, as
now we resize the image all the time, which is obviously bad. I also
wanted the image not to only fit the width but the height too, but that
also was a pain.