Create Slider component
There are no specs for this component, and no Swing implementation.
However, this has been derived from the general IntUI style with Chris'
help to choose behaviour, appearance and colours.

It is needed for a demo, hence why it was implemented already; if the
IntUI specs will eventually include this component, we'll update the
implementation as needed.