Enable publishing release-specific ui/foundation
Since the implementations change between IJP versions, we need to be
able to ship different versions of the dependency for different IJP
versions.

Since this only impacts the bridge, and the user already has to specify
the IJP-specific version of the bridge, and these are only transitive
dependencies, this should be a transparent change for users.