Fix text colour in Markdown (bridge)
We were not setting the right text colour when creating the bridge
Markdown Styling. Now we make sure we do respect the colour
set in the various InlinesStyle's TextStyles when rendering inlines.

In addition, I've added a simple API to be able to initialize a
MarkdownStyling in the bridge based only on a base TextStyle. This is
useful, e.g., when you want to make smaller text such as comments.