Add Markdown styling to theme, add simple components
We were lacking a good, easy way to render Markdown, requiring a lot of
ceremony. Now it's much easier, you just need to provide the styling,
renderer, and processor as composition locals (with the defaults
coming from ProvideMarkdownStyling). Then use either Markdown(), or
LazyMarkdown(), and your content is displayed. Better!

I have updated the samples to show how it works.