Add text colours to GlobalColors
It's practical to have the "base" text colours in GlobalColors:

 * Normal
 * Selected
 * Disabled
 * Info
 * Error

They apply to labels and most components, but not every component
(e.g., default buttons have "inverted" text colors).

Note that the text colour is used as content colour, but not directly
in the styling. Besides, when initialising the stylings, we can't
really read these values as they haven't been init yet.

This is mostly a convenience for users that want to set a content colour
(e.g., to visually show some text as disabled/error).

Note: GlobalColors.paneBackground was renamed to panelBackground for
consistency with the Swing color key naming.