Remove ktlint IJ plugin from required plugins
While it's nice to get ktlint warnings in the IDE, the ktlint IJ plugin
shows an obnoxious editor banner when working on projects where it's not
used.

Since we all work on other projects, and because their maintainers are
not open to improving the UX, I don't recommend keeping this plugin
anymore.