Fix Markdown rendering issues (#458)
* Make Markdown processing not optimised by default

The vast majority of Markdown use cases will be displaying static text,
so trying to optimise for an editor-like scenario is counter-productive.
This PR sets the flag to false by default.

This commit also cleans up the MarkdownProcessor code and kdocs.

* Make Markdown blocks implement equals and hashcode

The value classes we used to use would rely on the CommonMark models for
this, but none of the CommonMark models actually implement equals() and
hashCode(), leading to issues in tests and in Compose.

* Cleanup links in JewelReadme

Now links point to GitHub, so the app doesn't crash when you click them.

* Add equals/hashcode to inline Markdown entities, too

Tests have been updated to match the new behaviour, that's closer to the
CommonMark one. Now we don't carry unparsed inline Markdown anymore, but
rather we fully parse it in advance, so rendering it later is easier and
doesn't require running parsing in the UI anymore.

This is a tradeoff in memory usage for speed of rendering, and it makes
sense in the context of most Markdown text being static and only parsed
once, then kept on screen.

Inline extensions are now possible, but not tested. This makes #325
possible, since the extension now can be moved to its own module.

Note: please don't look too much at the test changes. They're _verbose_,
and are only there to ensure CommonMark specs compliance.

* Add inline extensions mechanism

This is not tested, and not used yet, so it may or may not need tweaks
down the line. We'll find out when we implement the first extension.