Add selectable and toggleable Icon(Action)Button
Introduces two variants for `IconButton` and `IconActionButton`:
* `SelectableIcon(Action)Button`
* `ToggleableIcon(Action)Button`

What they do should be self-evident.

Other changes include adding the ability to provide `PainterHint`s to
`IconActionButton`s, adding a standalone demo for the new icon buttons,
separating and cleaning out states and styling helpers.

Note: the main difference between an `IconButton` and its equivalent
`IconActionButton` is that the latter is built upon the former, and
handles automatically:
 * Tooltips (if defined)
 * Stroke modifiers (for selectable and toggleable)
 * Icon loading