Improve IconKey API to include a classloader
This prevents issues when running in release-mode IDEs, where providing
a classloader is mandatory to successfully load icons.

The icon keys generator script has also been updated to include the
class in each entry. This has been tested in standalone, in the normal
IDE sample mode, and as a built plugin in a release-mode IDE.