Change the New UI detection logic in the bridge (#675)
Somewhere in 242 or 243 the logic to determine whether a theme is "New
UI" in the IJP changed. It used to check the theme name, but these days
all themes are "New UI" themes when the New UI is enabled (which it is
now by default, as by 243 the old UI is an unbundled plugin).

So this changes the check we use to simply be NewUI.isEnabled().

Consider all themes "New UI" in bridge when the New UI is enabled

Even tho this feels weird, in IJP 243 (and possibly 242) all themes are
now considered to be "New UI" themes when the `NewUI.isEnabled()` flag
is enabled. This means that themes like High Contrast and Darcula are
also considered "New UI", despite it being unintuitive, at least in
Darcula's case (but High Contrast's parent is Darcula in the json, so...
ok it's weird that's it)