Markdown: synchronize scrolling between editor and preview (#690)
Only ScrollState is supported because of
its natural ability to scroll the view
to an arbitrary coordinate,

LazyListState doesn't allow this, and it
only gives an opportunity to scroll to
an item in a LazyColumn list and then
to a position within the item.
So, it requires a different approach
(which can hopefully be adjusted to the
proposed ScrollingSynchronizer API).

Note that the change only enables auto-scrolling
in a preview to match the position in the source,
it doesn't work the other way around.